[#146] extension 위치 조정
